+++ lib/ui/wallet/WalletLibraryCardPage.dart
@@ -38,6 +38,12 @@ class _WalletLibraryCardPageState extends State<WalletLibraryCardPage> with Noti
       Auth2.notifyCardChanged,
     ]);
 
+    // int daysBetween(DateTime from, DateTime to) {
+    //   from = DateTime(from.year, from.month, from.day);
+    //   to = DateTime(to.year, to.month, to.day);
+    //   return (to.difference(from).inHours / 24).round();
+    // }
+
     _barcodeNumber = Auth2().iCard?.libraryNumber;
     WidgetsBinding.instance.addPostFrameCallback((_) {
       _loadBarcodeImage(_barcodeNumber).then((MemoryImage? barcodeImage) {
@@ -94,7 +100,13 @@ class _WalletLibraryCardPageState extends State<WalletLibraryCardPage> with Noti
 
   @override
   Widget build(BuildContext context) {
-    return WalletPhotoWrapper(topOffset: widget.topOffset, headingColor: Styles().colors.lightGray, child: _buildCardContent());
+    return WalletPhotoWrapper(
+      topOffset: widget.topOffset,
+      headingColor: Styles().colors.lightGray,
+      accentColor: _photoBorderDay ? Styles().colors.libraryCardAccentBlue : Styles().colors.libraryCardAccentOrange,
+      borderColor: _photoBorderDay ? Styles().colors.libraryCardBorderBlue : Styles().colors.libraryCardBorderOrange,
+      child: _buildCardContent()
+    );
   }
 
   Widget _buildCardContent() {
@@ -160,6 +172,7 @@ class _WalletLibraryCardPageState extends State<WalletLibraryCardPage> with Noti
   String? get _displayAccessTime => AppDateTime().formatDateTime(_accessTime, format: 'MMM dd, yyyy HH:mm a');
   double get _barcodeWidth => MediaQuery.of(context).size.width / 1.25;
   double get _barcodeHeight => _barcodeWidth / 4;
+  bool get _photoBorderDay => ((((_accessTime ?? DateTime.now()).difference(DateTime(1970)).inHours / 24).round() % 2) == 0); // Used to alternate the photo border color scheme daily
 
   Widget get _barcodeImageWidget =>
     Container(width: _barcodeWidth, height: _barcodeHeight, decoration: BoxDecoration(